Many classes were modified.

+++ lecture-service/src/main/java/com/syrisa/lectureservice/utilities/generate/impl/NumberGenerate.java
@@ -5,12 +5,12 @@ import com.syrisa.lectureservice.utilities.generate.Generate;
 import java.util.Random;
 
 public class NumberGenerate {
-    public static Generate<Long> generate = numberSize -> {
+    public static Generate<Integer> generate = numberSize -> {
         StringBuilder builder = new StringBuilder();
         builder.append(1 + new Random().nextInt(9));
         for (int i = 0; i < numberSize - 1; i++) {
             builder.append(new Random().nextInt(10));
         }
-        return Long.parseLong(builder.toString());
+        return Integer.parseInt(builder.toString());
     };
 }
